setup linter/formatter and make other changes
Eslint and prettier had already been used for the project. I had
removed them some time before the first commit mainly because I was
having problems with them. I reinstalled them.

I felt like improving the function descriptions in the html table and
adding a column for the function arguments.

Upon review of the table, I noticed that the order of arguments for
getTasksWithTags was different than other getFns so I reversed the
order.

I also changed getTasksWithWords to getTasksWithText because it doesn't
only check for valid words, it works for gibberish as well.

I also changed the name of an argument from terms to strings because the
meaning of string is well known and I think I would cause someone
confusion if i were to call it something else. I dont have a good reason
for the initial name.

Finally, I added more notes below the table with the intent to help
someone understand my program.